% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_exprs_matrix}
\alias{as_exprs_matrix}
\title{Create an expression matrix.}
\usage{
as_exprs_matrix(data, container)
}
\arguments{
\item{data}{a data frame, column names should contain gene names or
identifiers.}

\item{container}{a dbsig database object or a list with the
gene name vectors.}
}
\value{
an expression matrix: gene names are presented in row names,
columns represent the samples.
}
\description{
Creates an matrix with the gene expression values in a format
accepted by the \code{\link[GSVA]{gsva}} function.
}
