% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{calculate}
\alias{calculate}
\title{Calculate object statistic.}
\usage{
calculate(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{arguments passed to the methods.}
}
\description{
Calculate object statistics.
}
