% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{dbsig}
\alias{dbsig}
\title{Create a dbsig object from a data tibble.}
\usage{
dbsig(x)
}
\arguments{
\item{x}{a tibble.}
}
\value{
an object of class dbsig.
}
\description{
Converts a data tibble into a dbsig object.
}
\details{
Technically, the object inherits most of the properties and methods
from the tibble and data.frame classes.
}
