% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_null_}
\alias{empty_null_}
\title{Filter out empty strings.}
\usage{
empty_null_(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
a character vector.
}
\description{
Removes empty strings from a character vector.
}
