% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{is_dbsig}
\alias{is_dbsig}
\title{Check for the dbsig class.}
\usage{
is_dbsig(x)
}
\arguments{
\item{x}{an object.}
}
\value{
a logical value.
}
\description{
Checks is an object is an instance of the dbsig class.
}
