% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_dbsig}
\alias{load_dbsig}
\title{Load a GSEA database.}
\usage{
load_dbsig(path)
}
\arguments{
\item{path}{a path to the .gmt file containing GSEA signatures.}
}
\value{
a dbsig object.
}
\description{
Loads a GSEA database from a local file and converts it
to a data frame of the dbsig class.
}
