% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{search.dbsig}
\alias{search.dbsig}
\title{Search in the dbsig database with a text expression.}
\usage{
\method{search}{dbsig}(x, key = c("sign_name", "genes", "sign_link"), value = ".*")
}
\arguments{
\item{x}{a dbsig object.}

\item{key}{the dbsig database field, see description. 'sign_name' by default.}

\item{value}{a regular expression used to filter the database.}
}
\value{
a dbsig object.
}
\description{
Filters entries of a dbsig database with a regular expression.
The fields available for searching are 'genes', 'sign_name'
(official signature name) and 'sign_link' (signature URL).
}
