% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{calculate.default}
\alias{calculate.default}
\alias{calculate.dbsig}
\title{Calculate GSVA scores.}
\usage{
\method{calculate}{default}(x, data, ...)

\method{calculate}{dbsig}(x, data, ...)
}
\arguments{
\item{x}{a list with gene name/identifier vectors or a dbsig object.}

\item{data}{a data frame with the expression values.}

\item{...}{extra arguments passed to \code{\link[GSVA]{gsva}}.}
}
\value{
A data frame with GSVA enrichment scores:
rows are samples, columns are scores.
}
\description{
Calculates multi-gene scores via \code{\link[GSVA]{gsva}}
function, given a character vector with gene names and a data frame with
expression values.
}
\details{
See: \code{\link[GSVA]{gsva}}.
}
