% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{calculate.default}
\alias{calculate.default}
\alias{calculate.dbsig}
\title{Calculate GSVA scores.}
\usage{
\method{calculate}{default}(
  x,
  data,
  id_col = "sample_id",
  BPPARAM = SerialParam(progressbar = TRUE),
  ...
)

\method{calculate}{dbsig}(x, data, id_col = "sample_id", ...)
}
\arguments{
\item{x}{a list with gene name/identifier vectors or a dbsig object.}

\item{data}{a data frame with the expression values.}

\item{id_col}{name of the column to store identifiers/names of the samples
in the output data frames.}

\item{BPPARAM}{An object of class BiocParallelParam
(https://www.bioconductor.org/packages//release/bioc/vignettes/BiocParallel/inst/doc/Introduction_To_BiocParallel.html)
with parameters of paralellization used by \code{\link[GSVA]{gsva}}.}

\item{...}{extra arguments passed to \code{\link[GSVA]{gsvaParam}}.}
}
\value{
A data frame with GSVA enrichment scores:
rows are samples, columns are scores.
}
\description{
Calculates multi-gene scores via \code{\link[GSVA]{gsva}}
function, given a character vector with gene names and a data frame with
expression values.
}
\details{
See: \code{\link[GSVA]{gsva}}.
}
