% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP.R
\name{search}
\alias{search}
\title{Search for a database entry.}
\usage{
search(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{arguments passed to the methods.}
}
\description{
Search for a database entry.
}
